added ratings to home page and products page
